# pkg/api/openapi.yaml
openapi: "3.0.3"
info:
  title: GoFast Machine API
  description: >-
    Machine-tier REST API served by pkg/api.NewRouter / RegisterMachineRoutes:
    a health check and read-only SQL access over parsed MySQL slow-query
    data. This document covers ONLY the machine tier (no dashboard/UI
    routes). info.version is the frozen machine-API contract version
    (api.ContractVersion), independent of the software release tag.
  version: "1.0.0"
paths:
...
  /api/v1/sql/databases:
    get:
      operationId: listDatabases
      summary: List available database names
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabasesResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    DatabasesResponse:
      type: object
      required: [databases]
      properties:
        databases:
          type: array
          nullable: false
          items:
            type: string
...
    ErrorResponse:
      type: object
      description: The uniform error envelope returned by every non-2xx response.
      required: [error, message]
      properties:
        error:
          type: string
          description: A stable machine-readable error code (e.g. invalid_token).
        message:
          type: string
          description: A human-readable message.
        request_id:
          type: string